move html content parser to util fn
