+++ src/v3/src/components/List/List.tsx
@@ -23,10 +23,9 @@ import {
   UISchemaElementComponent,
   UISchemaLayout,
 } from '../../types';
-import { getElementKey } from '../../util';
+import { getElementKey, parseHtmlContent } from '../../util';
 import Button from '../Button';
 import InformationalText from '../InformationalText';
-import ListItemText from './ListItemText';
 
 const renderElement = (item: UISchemaElement) => {
   const Container: FunctionComponent = ({ children }) => {
@@ -103,7 +102,7 @@ const List: UISchemaElementComponent<{
               // TODO: OKTA-577905 - (textAlign: 'start') Temporary fix until we can upgrade to the latest version of Odyssey
               sx={{ display: 'list-item', textAlign: 'start' }}
             >
-              {typeof item === 'string' ? (<ListItemText text={item} />) : renderLayout(item) }
+              {typeof item === 'string' ? parseHtmlContent(item) : renderLayout(item) }
             </ListItem>
           ))
         }
