--- src/v3/src/components/List/ListItemText.tsx
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2022-present, Okta, Inc. and/or its affiliates. All rights reserved.
- * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
- *
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *
- * See the License for the specific language governing permissions and limitations under the License.
- */
-
-import { Box } from '@okta/odyssey-react-mui';
-import { FunctionComponent, h } from 'preact';
-
-import { useHtmlContentParser } from '../../hooks';
-
-type ListItemTextProps = {
-  text: string;
-};
-
-const ListItemText: FunctionComponent<ListItemTextProps> = ({ text }) => {
-  const parsedText = useHtmlContentParser(text);
-
-  return (
-    <Box component="span">{ parsedText }</Box>
-  );
-};
-export default ListItemText;
