+++ src/v3/src/hooks/index.ts
@@ -12,7 +12,6 @@
 
 export * from './useAutoFocus';
 export * from './useFormFieldValidation';
-export * from './useHtmlContentParser';
 export * from './useInteractionCodeFlow';
 export * from './useOnce';
 export * from './useOnChange';
